requirements.txt
Flask==2.3.3
Flask-Caching==2.1.0
supabase==2.0.2
python-dotenv==1.0.0
Werkzeug==2.3.7 